Create MyFirstJob.java
Software Developer test
Purpose
The purpose of this test is to gauge the applicant’s proficiency in taking a specification and using it to create a real-world application.

Measurement
The results of the test will be judged in terms of whether or not the software performs the specified task as well as the thoughtfulness of the approach taken to solve the problem.

Deliverables
The applicant is required to hand in a compiled program that performs the specified task.  The applicant is also required to hand in all sources code necessary to create the compiled program.

Requirements
A program must be developed in Java or C#.
The program must perform the following tasks:
1.	Read the attached CSV file from the same folder as the executable program
2.	Create an XML file containing the contents of the CSV file.
3.	Display the contents of the CSV file on the screen sorted by AMOUNT.
4.	Display the total of the field AMOUNT.

Additional
The program can be written in Java or C# (Free tools are available for either on the web).
The program can be written as a Windows or Console app.
No user interface controls are necessary but can be added at your discretion.